+++ lib/presentation/screen/setting_screen/setting_screen.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import 'package:food_quest/domain/notifier/auth_notifier.dart';
+import 'package:food_quest/presentation/component/button.dart';
+
+class SettingScreen extends HookConsumerWidget {
+  const SettingScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final authNotifier = ref.watch(authNotifierProvider.notifier);
+
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            const Text('設定画面'),
+            CustomButton(
+              text: 'ログアウト',
+              onPressed: () async {
+                await authNotifier.logout();
+              },
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
